+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/universal/ScopedSecureStoreModule.kt
@@ -9,8 +9,8 @@ import host.exp.exponent.Constants
 
 private const val SHARED_PREFERENCES_NAME = "SecureStore"
 
-class ScopedSecureStoreModule(private val scopedContext: ScopedContext) :
-  SecureStoreModule(if (Constants.isStandaloneApp()) scopedContext.baseContext else scopedContext) {
+class ScopedSecureStoreModule(private val scopedContext: ScopedContext) : SecureStoreModule() {
+  override var reactContext: Context = if (Constants.isStandaloneApp()) scopedContext.baseContext else scopedContext
 
   // In standalone apps on SDK 41 and below, SecureStore was initiated with scoped context,
   // so SharedPreferences was scoped to that particular experienceId. This meant you