+++ packages/expo-secure-store/android/src/main/java/expo/modules/securestore/SecureStoreExceptions.kt
@@ -1,27 +1,24 @@
 package expo.modules.securestore
 
-import expo.modules.core.errors.CodedException
+import expo.modules.kotlin.exception.CodedException
 
 internal class NullKeyException :
   CodedException("SecureStore keys must not be null")
 
-internal class WriteException(message: String?, cause: Throwable?) :
-  CodedException(message ?: "An unexpected error occurred when writing to SecureStore", cause)
+internal class WriteException(message: String?, key: String, keychain: String, cause: Throwable? = null) :
+  CodedException("An error occurred when writing to key: '$key' under keychain: '$keychain'. Caused by: ${message ?: "unknown"}", cause)
 
-internal class ReadException(cause: Throwable?) :
-  CodedException("An unexpected error occurred when reading from SecureStore", cause)
+internal class EncryptException(message: String?, key: String, keychain: String, cause: Throwable? = null) :
+  CodedException("Could not encrypt the value for key '$key' under keychain '$keychain'. Caused by: ${message ?: "unknown"}", cause)
 
-internal class SecureStoreIOException(cause: Throwable?) :
-  CodedException("There was an I/O error loading the keystore for SecureStore", cause)
+internal class DecryptException(message: String?, key: String, keychain: String, cause: Throwable? = null) :
+  CodedException("Could not decrypt the value for key '$key' under keychain '$keychain'. Caused by: ${message ?: "unknown"}", cause)
 
-internal class EncryptException(message: String?, cause: Throwable?) :
-  CodedException(message ?: "Could not encrypt the value for SecureStore", cause)
+internal class DeleteException(message: String?, key: String, keychain: String, cause: Throwable? = null) :
+  CodedException("Could not delete the value for key '$key' under keychain '$keychain'. Caused by: ${message ?: "unknown"}", cause)
 
-internal class DecryptException(message: String?, cause: Throwable?) :
-  CodedException(message ?: "Could not decrypt the value for SecureStore", cause)
+internal class AuthenticationException(message: String?, cause: Throwable? = null) :
+  CodedException("Could not Authenticate the user: ${message ?: "unknown"}", cause)
 
-internal class SecureStoreJSONException(message: String?, cause: Throwable?) :
-  CodedException(message, cause)
-
-internal class DeleteException(message: String?, cause: Throwable?) :
-  CodedException(message ?: "An unexpected error occurred when deleting from SecureStore", cause)
+internal class KeyStoreException(message: String?) :
+  CodedException("An error occurred when accessing the keystore: ${message ?: "unknown"}")