+++ packages/expo-secure-store/CHANGELOG.md
@@ -10,6 +10,10 @@
 
 ### 💡 Others
 
+## 12.8.1 - 2023-12-19
+
+_This version does not introduce any user-facing changes._
+
 ## 12.8.0 — 2023-12-12
 
 ### 🎉 New features